chore: refactor scheduler / task
