+++ client-engine/task/task_set_event_logging.go
@@ -122,5 +122,5 @@ func LogTasksetEvent(t *TaskSet) {
 	eventLogJsonString, _ := json.Marshal(tasksetJson)
 	event_group_id := fmt.Sprintf("%s-%s", t.TaskSetId, t.Session.SessionId)
 	eventLog := NewEventLog("taskset", string(eventLogJsonString), event_group_id)
-	t.Session.ResourceTracker.ResourceManager.LogEvent(eventLog)
+	t.Session.NodeScheduler.ResourceManager.LogEvent(eventLog)
 }
